+++ src/main/java/web/service/UserService.java
@@ -7,7 +7,7 @@ import java.util.List;
 public interface UserService {
     void addUser(User user);
     List<User> getListUsers();
-    User show(int id);
-    void update(int id, User user);
-    void remove(int id);
+    User show(long id);
+    void update(long id, User user);
+    void remove(long id);
 }
